Added and set launcher icon for the app.
Set default orientation to portrait in Android Manifest.

Limited the number of characters that can be entered for user's name and restricted it to a single line in Create User view.

Changed layout of widgets so it changes accordingly with different phone sizes.

I think I fixed ISSUE 1 but need to do more testing. It might have been an issue with clicking on the Go Back button without passing data using Intents.

ISSUE 4 FIXED: Disabled Start button in Workout Session after it is clicked once in order to prevent misclicks or user trying to press it more than once.

ISSUE 1: Errors pop up when clicking the Stop button in Workout Session screen and when loading the Results screen.

ISSUE 2: In WorkoutSession view, timer doesn't work properly as it skips over the 0 second mark as it counts down.

ISSUE 3: Getting incorrect results for total workout time, calories burned, and BMI in Results view.