Added comments to AndroidManifest.xml
Fixed the formulas for total calories burned and BMI and the resulting values are now rounded to 2 decimal places.

Removed unnecessary comments.

Added workout description for each workout type to Workout Session screen.

Implemented database with two tables.

Added and set launcher icon for the app.

Set default orientation to portrait in Android Manifest.

Limited the number of characters that can be entered for user's name and restricted it to a single line in Create User view.

Changed layout of widgets so it changes accordingly with different phone sizes.

ISSUE 1 FIXED: Errors pop up when clicking the Stop button in Workout Session screen and when loading the Results screen. It might have been an issue with clicking on the Go Back button without passing data using Intents.

ISSUE 2: In WorkoutSession view, timer doesn't work properly as it skips over the 0 second mark as it counts down.

ISSUE 3 FIXED: Getting incorrect results for total workout time, calories burned, and BMI in Results view.

ISSUE 4 FIXED: Disabled Start button in Workout Session after it is clicked once in order to prevent misclicks or user trying to press it more than once.